build: set up to use feature flags on deployment
[#OCD-2745]

[#OCD-2820]
