+++ package.json
@@ -4,12 +4,10 @@
   "license": "SEE LICENSE IN LICENSE",
   "repository": "chpladmin/chpl-website",
   "scripts": {
-    "build": "webpack --config webpack.prod.js",
-    "build:uat": "webpack --config webpack.dev-prod.js",
+    "build": "webpack",
     "html-lint": "htmllint src/app/**/*.html",
     "html-lint:verbose": "htmllint --rc .htmllintrc.verbose src/app/**/*.html",
-    "start": "webpack-dev-server --config webpack.dev.js",
-    "start:dist": "webpack-dev-server --config webpack.dev-prod.js",
+    "start": "webpack-dev-server --server",
     "profile": "webpack --profile --json > test_reports/stats.json",
     "test": "karma start --single-run",
     "test:auto": "karma start",
