--- webpack.dev.js
@@ -1,40 +1,0 @@
-'use strict';
-
-const webpack = require('webpack');
-const path = require('path');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const merge = require('webpack-merge');
-
-const base = require('./webpack.ci.js');
-
-module.exports = merge(base, {
-    devServer: {
-        disableHostCheck: true,
-        port: 3000,
-        proxy: {
-            '/rest': {
-                target: 'http://localhost:8181/chpl-service',
-                pathRewrite: {'^/rest' : ''},
-            },
-        },
-        watchOptions: {
-            ignored: '**/.*',
-        },
-    },
-    plugins: [
-        new HtmlWebpackPlugin({
-            filename: 'style.html',
-            hash: true,
-            inject: 'body',
-            template: path.resolve(__dirname, './src/style.html'),
-        }),
-        new webpack.DefinePlugin({
-            DEVELOPER_MODE: JSON.stringify(true),
-            ENABLE_LOGGING: JSON.stringify(true),
-            FEATURE_FLAGS: JSON.stringify(require('./flags.dev.json')),
-            MINUTES_UNTIL_IDLE: 150,
-            MINUTES_BETWEEN_KEEPALIVE: 1,
-            UAT_MODE: JSON.stringify(true),
-        }),
-    ]
-});
